+++ extensions/pyRevitDevTools.extension/pyRevitDev.tab/Debug.panel/Misc Tests.pulldown/Test Rhino3dmIO.pushbutton/script.py
@@ -1,13 +1,13 @@
-from pyrevit import rhino as rc
+from winterops.rhino import Rhino
 
 
-pt1 = rc.Geometry.Point3d(0, 0, 0)
-pt2 = rc.Geometry.Point3d(10, 10, 0)
-line1 = rc.Geometry.Line(pt1, pt2)
+pt1 = Rhino.Geometry.Point3d(0, 0, 0)
+pt2 = Rhino.Geometry.Point3d(10, 10, 0)
+line1 = Rhino.Geometry.Line(pt1, pt2)
 line1.Length
-pt1 = rc.Geometry.Point3d(10, 0, 0)
-pt2 = rc.Geometry.Point3d(0, 10, 0)
-line2 = rc.Geometry.Line(pt1, pt2)
+pt1 = Rhino.Geometry.Point3d(10, 0, 0)
+pt2 = Rhino.Geometry.Point3d(0, 10, 0)
+line2 = Rhino.Geometry.Line(pt1, pt2)
 
 
-print(rc.Geometry.Intersect.Intersection.LineLine(line1, line2))
+print(Rhino.Geometry.Intersect.Intersection.LineLine(line1, line2))
