--- pyrevitlib/vstore/__init__.py
@@ -1,7 +1,0 @@
-import vstore.bbx
-import vstore.pts
-import vstore.stl
-import vstore.xl
-
-
-__version__ = '0.1'
