+++ pyrevitlib/winterops/dxf.py
@@ -1,0 +1,6 @@
+# pylama:ignore=E402,W0611
+from winterops import clr, System
+
+clr.AddReference('IxMilia.Dxf')
+
+from IxMilia import Dxf
