+++ pyrevitlib/winterops/tests/__init__.py
@@ -1,0 +1,3 @@
+import os.path as op
+
+USER_DESKTOP = op.join(op.expandvars('%userprofile%\\desktop'))
